Removed unnecessary file (#63)
* attempt to tif bugs in the verse check default area

* REMOVED unused code
